[ITB-1366] Steps within groups and loops not included in test session reports

+++ gitb-ui/app/controllers/RepositoryService.scala
@@ -229,7 +229,7 @@ class RepositoryService @Inject() (implicit ec: ExecutionContext, authorizedActi
       }
     } catch {
       case e: Exception =>
-        if (result._2.archived) {
+        if (result != null && result._2.archived) {
           FileUtils.deleteQuietly(result._2.path.toFile)
         }
         throw e
